LayoutTests/imported/w3c/web-platform-tests/streams/idlharness-shadowrealm.window-expected.txt


PASS ReadableStreamDefaultReader includes ReadableStreamGenericReader: member names are unique
PASS ReadableStreamBYOBReader includes ReadableStreamGenericReader: member names are unique
PASS ReadableStream interface: existence and properties of interface object
PASS ReadableStream interface object length
PASS ReadableStream interface object name
PASS ReadableStream interface: existence and properties of interface prototype object
PASS ReadableStream interface: existence and properties of interface prototype object's "constructor" property
PASS ReadableStream interface: existence and properties of interface prototype object's @@unscopables property
FAIL ReadableStream interface: attribute locked assert_true: property should be enumerable expected true got false
FAIL ReadableStream interface: operation cancel(optional any) assert_true: property should be enumerable expected true got false
FAIL ReadableStream interface: operation getReader(optional ReadableStreamGetReaderOptions) assert_true: property should be enumerable expected true got false
FAIL ReadableStream interface: operation pipeThrough(ReadableWritablePair, optional StreamPipeOptions) assert_true: property should be enumerable expected true got false
FAIL ReadableStream interface: operation pipeTo(WritableStream, optional StreamPipeOptions) assert_true: property should be enumerable expected true got false
FAIL ReadableStream interface: operation tee() assert_true: property should be enumerable expected true got false
FAIL ReadableStream interface: async iterable<any> undefined is not an object (evaluating 'desc.value')
PASS ReadableStreamDefaultReader interface: existence and properties of interface object
PASS ReadableStreamDefaultReader interface object length
PASS ReadableStreamDefaultReader interface object name
PASS ReadableStreamDefaultReader interface: existence and properties of interface prototype object
PASS ReadableStreamDefaultReader interface: existence and properties of interface prototype object's "constructor" property
PASS ReadableStreamDefaultReader interface: existence and properties of interface prototype object's @@unscopables property
PASS ReadableStreamDefaultReader interface: operation read()
PASS ReadableStreamDefaultReader interface: operation releaseLock()
PASS ReadableStreamDefaultReader interface: attribute closed
PASS ReadableStreamDefaultReader interface: operation cancel(optional any)
FAIL ReadableStreamBYOBReader interface: existence and properties of interface object assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface object length assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface object name assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: existence and properties of interface prototype object assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: operation read(ArrayBufferView) assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: operation releaseLock() assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: attribute closed assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: operation cancel(optional any) assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
PASS ReadableStreamDefaultController interface: existence and properties of interface object
PASS ReadableStreamDefaultController interface object length
PASS ReadableStreamDefaultController interface object name
PASS ReadableStreamDefaultController interface: existence and properties of interface prototype object
PASS ReadableStreamDefaultController interface: existence and properties of interface prototype object's "constructor" property
PASS ReadableStreamDefaultController interface: existence and properties of interface prototype object's @@unscopables property
PASS ReadableStreamDefaultController interface: attribute desiredSize
PASS ReadableStreamDefaultController interface: operation close()
PASS ReadableStreamDefaultController interface: operation enqueue(optional any)
PASS ReadableStreamDefaultController interface: operation error(optional any)
FAIL ReadableByteStreamController interface: existence and properties of interface object assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface object length assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface object name assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: existence and properties of interface prototype object assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: attribute byobRequest assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: attribute desiredSize assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: operation close() assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: operation enqueue(ArrayBufferView) assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: operation error(optional any) assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableStreamBYOBRequest interface: existence and properties of interface object assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface object length assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface object name assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface: existence and properties of interface prototype object assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface: attribute view assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface: operation respond(unsigned long long) assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface: operation respondWithNewView(ArrayBufferView) assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
PASS WritableStream interface: existence and properties of interface object
PASS WritableStream interface object length
PASS WritableStream interface object name
PASS WritableStream interface: existence and properties of interface prototype object
PASS WritableStream interface: existence and properties of interface prototype object's "constructor" property
PASS WritableStream interface: existence and properties of interface prototype object's @@unscopables property
PASS WritableStream interface: attribute locked
PASS WritableStream interface: operation abort(optional any)
PASS WritableStream interface: operation close()
PASS WritableStream interface: operation getWriter()
PASS WritableStreamDefaultWriter interface: existence and properties of interface object
PASS WritableStreamDefaultWriter interface object length
PASS WritableStreamDefaultWriter interface object name
PASS WritableStreamDefaultWriter interface: existence and properties of interface prototype object
PASS WritableStreamDefaultWriter interface: existence and properties of interface prototype object's "constructor" property
PASS WritableStreamDefaultWriter interface: existence and properties of interface prototype object's @@unscopables property
PASS WritableStreamDefaultWriter interface: attribute closed
PASS WritableStreamDefaultWriter interface: attribute desiredSize
PASS WritableStreamDefaultWriter interface: attribute ready
PASS WritableStreamDefaultWriter interface: operation abort(optional any)
PASS WritableStreamDefaultWriter interface: operation close()
PASS WritableStreamDefaultWriter interface: operation releaseLock()
PASS WritableStreamDefaultWriter interface: operation write(optional any)
PASS WritableStreamDefaultController interface: existence and properties of interface object
PASS WritableStreamDefaultController interface object length
PASS WritableStreamDefaultController interface object name
PASS WritableStreamDefaultController interface: existence and properties of interface prototype object
PASS WritableStreamDefaultController interface: existence and properties of interface prototype object's "constructor" property
PASS WritableStreamDefaultController interface: existence and properties of interface prototype object's @@unscopables property
PASS WritableStreamDefaultController interface: attribute signal
FAIL WritableStreamDefaultController interface: operation error(optional any) assert_equals: property has wrong .length expected 0 but got 1
PASS TransformStream interface: existence and properties of interface object
PASS TransformStream interface object length
PASS TransformStream interface object name
PASS TransformStream interface: existence and properties of interface prototype object
PASS TransformStream interface: existence and properties of interface prototype object's "constructor" property
PASS TransformStream interface: existence and properties of interface prototype object's @@unscopables property
PASS TransformStream interface: attribute readable
FAIL TransformStream interface: attribute writable assert_equals: getter must have the name 'get writable' expected "get writable" but got "writable"
PASS TransformStreamDefaultController interface: existence and properties of interface object
PASS TransformStreamDefaultController interface object length
PASS TransformStreamDefaultController interface object name
PASS TransformStreamDefaultController interface: existence and properties of interface prototype object
PASS TransformStreamDefaultController interface: existence and properties of interface prototype object's "constructor" property
PASS TransformStreamDefaultController interface: existence and properties of interface prototype object's @@unscopables property
PASS TransformStreamDefaultController interface: attribute desiredSize
PASS TransformStreamDefaultController interface: operation enqueue(optional any)
PASS TransformStreamDefaultController interface: operation error(optional any)
PASS TransformStreamDefaultController interface: operation terminate()
PASS ByteLengthQueuingStrategy interface: existence and properties of interface object
PASS ByteLengthQueuingStrategy interface object length
PASS ByteLengthQueuingStrategy interface object name
PASS ByteLengthQueuingStrategy interface: existence and properties of interface prototype object
PASS ByteLengthQueuingStrategy interface: existence and properties of interface prototype object's "constructor" property
PASS ByteLengthQueuingStrategy interface: existence and properties of interface prototype object's @@unscopables property
PASS ByteLengthQueuingStrategy interface: attribute highWaterMark
FAIL ByteLengthQueuingStrategy interface: attribute size assert_throws_js: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw
PASS CountQueuingStrategy interface: existence and properties of interface object
PASS CountQueuingStrategy interface object length
PASS CountQueuingStrategy interface object name
PASS CountQueuingStrategy interface: existence and properties of interface prototype object
PASS CountQueuingStrategy interface: existence and properties of interface prototype object's "constructor" property
PASS CountQueuingStrategy interface: existence and properties of interface prototype object's @@unscopables property
PASS CountQueuingStrategy interface: attribute highWaterMark
FAIL CountQueuingStrategy interface: attribute size assert_throws_js: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw
